pyproject.toml
[tool.poetry]
name = "grade_project"
version = "0.1.4"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.104.1"
uvicorn = "^0.27.0.post1"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
bs4 = "^0.0.2"
sqlalchemy = "^2.0.25"
pyjwt = "^2.8.0"
psycopg2-binary = "^2.9.9"
json-stream = "^2.3.2"
python-slugify = "^8.0.2"
alembic = "^1.13.1"
python-multipart = "^0.0.6"
boto3 = "^1.34.32"
pydantic-extra-types = "^2.5.0"
requests = "^2.31.0"
python-ulid = "^2.2.0"
lxml = "^5.1.0"
httpx = "^0.27.0"
fastapi-cache2 = "^0.2.1"
redis = "^5.0.3"
aioredis = "^2.0.1"
pytest = "^8.1.1"
aioboto3 = "^12.4.0"
lxml-html-clean = "^0.1.1"
sqladmin = "0.16.1"
fastapi-pagination = "^0.12.26"
sqlakeyset = "^2.0.1724199169"
celery = "^5.4.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
jinja2 = "^3.1.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
asgi-sitemaps = "==1.*"
pillow = "^10.4.0"
fastapi-mail = "^1.4.1"
fastapi-localization = "^0.0a2.dev3"
build = "^1.2.2"
i18n = "^0.2"
python-i18n = "^0.3.9"
fastapi-validation-i18n = "^0.4.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.1"
ipython = "^8.21.0"

[tool.ruff]
line-length = 120
target-version = "py310"
fix = false
show-fixes = true
output-format = "grouped"
extend-exclude = ["*.pyi"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = [
  "A",      # flake8-builtins
  "ANN",    # flake8-annotations
  "ARG",    # flake8-unused-arguments
  "ASYNC",  # flake8-async
  "B",      # flake8-bugbear
  "C4",     # flake8-comprehensions
  "C90",    # mccabe
  "DTZ",    # flake8-datetimez
  "E",      # pycodestyl
  "EM",     # flake8-errmsg
  "ERA",    # eradicate
  "F",      # pyflakes
  "FBT",    # flake8-boolean-trap
  "I",      # isort
  "INT",    # flake8-gettext
  "N",      # pep8-naming
  "PL",     # Pylint
  "PERF",   # Perflint
  "PT",     # flake8-pytest-style
  "RET",    # flake8-return
  "RUF",    # Ruff-specific rules
  "S",      # flake8-bandit
  "SLF",    # flake8-flake8-self
  "T20",    # flake8-print
  "TRY",    # tryceratops
  "SIM",    # flake8-simplify
  "UP",     # pyupgrade
  "W",      # pycodestyle
]
ignore = ["ANN101", "RUF001", "RUF003"]
ignore-init-module-imports = true

[tool.ruff.lint.extend-per-file-ignores]
"**/{tests,docs,tools}/*" = ["E402", "ANN401", "ANN001", "S603", "ARG001", "S101", "S311"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.isort]
lines-between-types = 1
lines-after-imports = 2

[tool.ruff.lint.mccabe]
# Flag errors (C901) whenever the complexity level exceeds 6.
max-complexity = 6

[tool.ruff.lint.pylint]
allow-dunder-method-names = ["__tablename__", "__table_args__", "__post_init__"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
console_output_style = "count"
norecursedirs = "docker"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
